+++ app/src/main/java/de/triology/universeadm/user/imports/ImportError.java
@@ -1,9 +1,6 @@
 package de.triology.universeadm.user.imports;
 
 import com.google.common.collect.ImmutableMap;
-import org.omg.CORBA.INTERNAL;
-
-
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;