+++ app/src/main/ui/src/hooks/useAPI.ts
@@ -1,14 +1,21 @@
-import { useEffect, useState} from "react";
+import {useEffect, useState} from "react";
 
-export type QueryOptions = {start: number; limit: number; query: string; exclude?: string[];}
+export type QueryOptions = { start: number; limit: number; query: string; exclude?: string[]; }
 
 export type AbortableCallbackWithArgs<T, Y> = (_signal?: AbortSignal, _args?: Y) => Promise<T>
 export type AbortableCallback<T> = (_signal?: AbortSignal) => Promise<T>
 
-export function useAPI<T, Y>(_callBack: AbortableCallbackWithArgs<T, Y>, _args?: Y): any;
-export function useAPI<T>(_callBack: AbortableCallback<T>): any;
+export interface UseApiResponse<T> {
+    data: T | undefined,
+    setData: (_: T) => void,
+    isLoading: boolean,
+    error: any
+}
 
-export function useAPI<T,Y>(callBack: AbortableCallbackWithArgs<T, Y> | AbortableCallback<T>, args?: Y) {
+export function useAPI<T, Y>(_callBack: AbortableCallbackWithArgs<T, Y>, _args?: Y): UseApiResponse<T>;
+export function useAPI<T>(_callBack: AbortableCallback<T>): UseApiResponse<T>;
+
+export function useAPI<T, Y>(callBack: AbortableCallbackWithArgs<T, Y> | AbortableCallback<T>, args?: Y): UseApiResponse<T> {
     const [data, setData] = useState<T>();
     const [isLoading, setIsLoading] = useState<boolean>(true);
     const [error, setError] = useState<Error>();
@@ -21,7 +28,7 @@ export function useAPI<T,Y>(callBack: AbortableCallbackWithArgs<T, Y> | Abortabl
                 setIsLoading(false);
             })
             .catch(err => {
-                if(!abortController.signal.aborted) {
+                if (!abortController.signal.aborted) {
                     setError(err);
                     console.error(err);
                 }